refactor(lowy): extract HEAD watcher singleton into head-watcher.ts
The refcounted singleton (Map, install helper, watchGitHead, test
inspector) is its own volatility cluster — debounce strategy and
event-source policy change independently of git-info resolution.
Co-locating with `resolveGitInfo`/`hasGitDir` meant any of those
axes touched the same file. Lifts to a sibling module the way
`wal-subscription.ts` lives next to its consumers.
